pyproject.toml
[tool.poetry]
name = "yamiconfig"
version = "0.1.0"
description = "Yet another application configuration library"
authors = ["Tim McFadden <tim@timandjamie.com>"]

[tool.poetry.dependencies]
python = "^3.7"
"ruamel.yaml" = "^0.16.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pylint = "^2.4.4"
ipdb = "^0.12.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
